requirement.txt
langchain
langchain-community
langchain-core
langchain_openai
unstructured
"unstructured[pdf]"
mysql-connector-python
streamlit